--- Authenticator/Authenticator/PingOne.framework/Modules/PingOne.swiftmodule/armv7-apple-ios.swiftinterface
@@ -1,81 +1,0 @@
-// swift-interface-format-version: 1.0
-// swift-compiler-version: Apple Swift version 5.1.3 (swiftlang-1100.0.282.1 clang-1100.0.33.15)
-// swift-module-flags: -target armv7-apple-ios10.0 -enable-objc-interop -enable-library-evolution -swift-version 5 -enforce-exclusivity=checked -Onone -module-name PingOne
-import AVKit
-import CommonCrypto
-import Compression
-import CoreFoundation
-import CoreLocation
-import Foundation
-import LocalAuthentication
-@_exported import PingOne
-import Security
-import Swift
-import SystemConfiguration
-import UIKit
-import UserNotifications
-@objc public enum ErrorCode : ObjectiveC.NSInteger {
-  case internalError = 10000
-  case deviceTokenIsMissing = 10001
-  case unrecognizedRemoteNotification = 10002
-  case serverError = 10003
-  case noConnectivity = 10004
-  case pairingKey = 10005
-  case bundleId = 10006
-  case pairingKeyDataCenterMismatch = 10007
-  case deviceIsNotPaired = 10008
-  case pushConfirmationTimeout = 10009
-  public typealias RawValue = ObjectiveC.NSInteger
-  public init?(rawValue: ObjectiveC.NSInteger)
-  public var rawValue: ObjectiveC.NSInteger {
-    get
-  }
-}
-@objc public class NotificationObject : ObjectiveC.NSObject {
-  @objc public enum NotificationType : ObjectiveC.NSInteger {
-    case none = -1
-    case done
-    case authentication
-    public typealias RawValue = ObjectiveC.NSInteger
-    public init?(rawValue: ObjectiveC.NSInteger)
-    public var rawValue: ObjectiveC.NSInteger {
-      get
-    }
-  }
-  @objc final public let notificationType: PingOne.NotificationObject.NotificationType
-  @objc final public let timeoutDuration: Swift.Int
-  @objc final public func approve(withAuthenticationMethod: Swift.String?, completionHandler: @escaping (Foundation.NSError?) -> Swift.Void)
-  @available(*, deprecated, message: "please use approve(withAuthenticationMethod:completionHandler) instead")
-  @objc final public func approve(completionHandler: @escaping (Foundation.NSError?) -> Swift.Void)
-  @objc final public func deny(completionHandler: @escaping (Foundation.NSError?) -> Swift.Void)
-  @objc deinit
-}
-@objc public class PairingObject : ObjectiveC.NSObject {
-  @objc final public func approve(completionHandler: @escaping (Foundation.NSError?) -> Swift.Void)
-  @objc deinit
-}
-@objc public class PingOne : ObjectiveC.NSObject {
-  @objc public enum APNSDeviceTokenType : ObjectiveC.NSInteger {
-    case production = 0
-    case sandbox
-    public typealias RawValue = ObjectiveC.NSInteger
-    public init?(rawValue: ObjectiveC.NSInteger)
-    public var rawValue: ObjectiveC.NSInteger {
-      get
-    }
-  }
-  @objc public static func generateMobilePayload() throws -> Swift.String
-  @objc public static func pair(_ pairingKey: Swift.String, completionHandler: @escaping (Foundation.NSError?) -> Swift.Void)
-  @objc public static func processIdToken(_ idToken: Swift.String, completionHandler: @escaping (PingOne.PairingObject?, Foundation.NSError?) -> Swift.Void)
-  @objc public static func setDeviceToken(_ deviceToken: Foundation.Data, type: PingOne.PingOne.APNSDeviceTokenType, completionHandler: @escaping (Foundation.NSError?) -> Swift.Void)
-  @objc public static func getUNNotificationCategories() -> Swift.Set<UserNotifications.UNNotificationCategory>
-  @objc public static func processRemoteNotification(_ userInfo: [Swift.AnyHashable : Any], completionHandler: @escaping (PingOne.NotificationObject?, Foundation.NSError?) -> Swift.Void)
-  @objc public static func processRemoteNotificationAction(_ identifier: Swift.String, authenticationMethod: Swift.String?, forRemoteNotification userInfo: [Swift.AnyHashable : Any], completionHandler: @escaping (PingOne.NotificationObject?, Foundation.NSError?) -> Swift.Void)
-  @objc public static func getInfo(_ completionHandler: @escaping ([Swift.String : Any]?, Foundation.NSError?) -> Swift.Void)
-  @available(*, deprecated, message: "please use processRemoteNotificationAction(identifier:authenticationMethod:userInfo) instead")
-  @objc public static func processRemoteNotificationAction(_ identifier: Swift.String, forRemoteNotification userInfo: [Swift.AnyHashable : Any], completionHandler: @escaping (PingOne.NotificationObject?, Foundation.NSError?) -> Swift.Void)
-  @objc public static func removePingOneLocalData()
-  @objc public static func setDevicePairedAfterReinstall(_ stayPaired: Swift.Bool)
-  @objc public static func sendLogs(completionHandler: @escaping (Swift.String?, Foundation.NSError?) -> Swift.Void)
-  @objc deinit
-}
