--- Authenticator/Authenticator/PingOne.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module PingOne {
-  umbrella header "PingOne.h"
-
-  export *
-  module * { export * }
-}
-
-module PingOne.Swift {
-    header "PingOne-Swift.h"
-    requires objc
-}
